backend/requirements.txt
Django==2.1.4
django-appconf==1.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-imagekit==4.0.2
django-import-export==1.1.0
django-ipware==2.1.0
django-redis-cache==1.8.1
django-rest-swagger==2.2.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.0
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.3.14
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.12
pilkit==2.0
Pillow==5.3.0
PyJWT==1.7.1
pytz==2018.7
PyYAML==3.13
redis==2.10.6
requests==2.21.0
simplejson==3.16.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
uWSGI==2.0.17.1
django-crequest==2018.5.11
python-ldap==3.1.0
ldap3==2.5.2
itsdangerous==1.1.0
social-auth-app-django==3.1.0
cryptography==2.6.1
sqlparse==0.3.0
django-celery==3.2.2
python-jenkins==1.4.0
pyotp==2.2.7
qrcode==6.1
python-gitlab==1.8.0
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
python-dotenv==0.10.3
aliyun-python-sdk-core==2.13.5
aliyun-python-sdk-ecs==4.16.11
aliyun-python-sdk-r-kvstore==2.1.1
tencentcloud-sdk-python==3.0.78
xpinyin==0.5.6
cs==2.7